+++ src/python/MoveToSomewhereElse/ModuleLock.py
@@ -8,7 +8,7 @@ from time import struct_time, gmtime, mktime, asctime, sleep
 
 from typing import Optional
 
-from Utils.Logging import displayTime
+from Utilities.Logging import displayTime
 from Services.Mongo.MongoClient import MongoClient
