+++ src/python/MoveToSomewhereElse/ReportInfo.py
@@ -3,7 +3,7 @@ from pymongo.collection import Collection
 
 from typing import Any
 
-from Utils.IteratorTools import mapKeys, mapValues
+from Utilities.IteratorTools import mapKeys, mapValues
 from Services.Mongo.MongoClient import MongoClient
